backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
python-dotenv
requests
openai
httpx
pandas
pdfplumber
python-docx
pypdf
pdfminer.six
python-multipart
openpyxl
xlrd
